requirements.txt
# Web Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0

# Database
sqlalchemy>=2.0.0
alembic>=1.12.0
psycopg2-binary>=2.9.0

# Cache & Session
redis>=5.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# AI Services
openai>=1.0.0
google-generativeai>=0.3.0
anthropic>=0.7.0

# Vector Database & RAG
supabase>=2.0.0
sentence-transformers>=2.2.0
faiss-cpu>=1.7.0
chromadb>=0.4.0

# Document Processing
pypdf2>=3.0.0
python-docx>=0.8.11
python-multipart>=0.0.6
markdown>=3.5.0
beautifulsoup4>=4.12.0

# HTTP Client & WebSocket
httpx>=0.25.0
aiohttp>=3.9.0
websockets>=12.0

# Utilities
python-dotenv>=1.0.0
pydantic-settings>=2.0.0
structlog>=23.2.0
typer>=0.9.0

# Development & Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
fakeredis>=2.20.0
black>=23.0.0
isort>=5.12.0
mypy>=1.7.0
pre-commit>=3.5.0

# Monitoring & Performance
prometheus-client>=0.19.0
psutil>=5.9.0

# Data Processing
pandas>=2.1.0
numpy>=1.24.0

# Security
cryptography>=41.0.0
bcrypt>=4.1.0

# CLI & Scripting
click>=8.1.0
rich>=13.0.0

# Email (Optional)
fastapi-mail>=1.4.0

# Image Processing (Optional)
Pillow>=10.0.0